Initial refactor for body support
