+++ src/addons/meta_human_dna/operators.py
@@ -5,11 +5,15 @@ import shutil
 import logging
 from pathlib import Path
 from datetime import datetime, timedelta
-from .face import MetahumanFace
 from .ui import importer, callbacks
 from . import utilities
 from .dna_io import DNACalibrator, DNAExporter, create_shape_key
 from .properties import MetahumanDnaImportProperties
+from .components import (
+    MetaHumanComponentHead,
+    MetaHumanComponentBody,
+    get_meta_human_component
+)
 from .constants import (
     SEND2UE_FACE_SETTINGS,
     TEXTURE_LOGIC_NODE_NAME,
@@ -65,12 +69,12 @@ class GenericProgressQueueOperator(bpy.types.Operator):
         
         self._timer = context.window_manager.event_timer_add(0.01, window=context.window) # type: ignore
         context.window_manager.modal_handler_add(self) # type: ignore
-        face = utilities.get_active_face()
-        if face:
+        head = utilities.get_active_head()
+        if head:
             context.window_manager.meta_human_dna.progress = 0 # type: ignore
             context.window_manager.meta_human_dna.progress_description = '' # type: ignore
             self._commands_queue = queue.Queue()
-            self.set_commands_queue(context, face, self._commands_queue)
+            self.set_commands_queue(context, head, self._commands_queue)
             self._commands_queue_size = self._commands_queue.qsize()
             return {'RUNNING_MODAL'}
         return {'CANCELLED'}
@@ -93,7 +97,7 @@ class GenericProgressQueueOperator(bpy.types.Operator):
     def set_commands_queue(
             self, 
             context, 
-            face: MetahumanFace,
+            component: MetaHumanComponentHead | MetaHumanComponentBody,
             commands_queue: queue.Queue
         ):
         pass   
@@ -113,9 +117,9 @@ class ImportAnimation(bpy.types.Operator, importer.ImportAsset):
 
     def execute(self, context):
         logger.info(f'Importing animation {self.filepath}')  # type: ignore
-        face = utilities.get_active_face()
-        if face:
-            face.import_action(Path(self.filepath))  # type: ignore
+        head = utilities.get_active_head()
+        if head:
+            head.import_action(Path(self.filepath))  # type: ignore
         return {'FINISHED'}
     
 class BakeAnimation(bpy.types.Operator):
@@ -192,6 +196,9 @@ class BakeAnimation(bpy.types.Operator):
         return context.window_manager.invoke_props_dialog(self, width = 250) # type: ignore
 
     def draw(self, context):
+        if not self.layout:
+            return
+        
         row = self.layout.row()
         row.prop(self, 'action_name', text="Name")
         row = self.layout.row()
@@ -217,8 +224,8 @@ class BakeAnimation(bpy.types.Operator):
             self.report({'ERROR'}, 'The start frame must be less than the end frame')
             return {'CANCELLED'}
 
-        face = utilities.get_active_face()
-        if face and face.head_rig_object:
+        head = utilities.get_active_head()
+        if head and head.head_rig_object:
             channel_types = set()
             if self.bone_location:
                 channel_types.add('LOCATION')
@@ -228,7 +235,7 @@ class BakeAnimation(bpy.types.Operator):
                 channel_types.add('SCALE')
 
             utilities.bake_to_action(
-                armature_object=face.head_rig_object,
+                armature_object=head.head_rig_object,
                 action_name=self.action_name,
                 start_frame=self.start_frame, # type: ignore
                 end_frame=self.end_frame, # type: ignore
@@ -293,11 +300,11 @@ class ImportMetahumanDna(bpy.types.Operator, importer.ImportAsset, MetahumanDnaI
 
         # we don't want to evaluate the dependency graph while importing the DNA
         window_manager_properties.evaluate_dependency_graph = False
-        face = MetahumanFace(
-            dna_file_path=file_path,
-            dna_import_properties=self.properties # type: ignore
+        component = get_meta_human_component(
+            file_path=file_path,
+            properties=self.properties # type: ignore
         )
-        valid, message = face.ingest()
+        valid, message = component.ingest()
         # populate the output items based on what was imported
         callbacks.update_output_items(None, bpy.context)
         logger.info(f'Finished importing "{self.filepath}"') # type: ignore
@@ -423,7 +430,7 @@ class ConvertSelectedToDna(bpy.types.Operator, MetahumanDnaImportProperties):
 
         # we don't want to evaluate the dependency graph while importing the DNA
         window_manager_properties.evaluate_dependency_graph = False
-        face = MetahumanFace(
+        face = MetaHumanComponentHead(
             name=new_name,
             dna_file_path=Path(self.base_dna),
             dna_import_properties=self.properties # type: ignore
@@ -555,9 +562,9 @@ class GenerateMaterial(bpy.types.Operator):
     bl_label = "Generate Material"    
 
     def execute(self, context):
-        face = utilities.get_active_face()
-        if face and face.head_mesh_object:
-            face.import_materials()
+        head = utilities.get_active_head()
+        if head and head.head_mesh_object:
+            head.import_materials()
         return {'FINISHED'}
     
     @classmethod
@@ -579,10 +586,10 @@ class ImportShapeKeys(GenericProgressQueueOperator):
     def set_commands_queue(
             self, 
             context, 
-            face: MetahumanFace,
+            component: MetaHumanComponentHead,
             commands_queue: queue.Queue
         ):
-        face.import_shape_keys(commands_queue)
+        component.import_shape_keys(commands_queue)
         bpy.ops.meta_human_dna.force_evaluate() # type: ignore
 
 
@@ -653,9 +660,9 @@ class SendToUnreal(bpy.types.Operator):
             logger.error('The Send to Unreal addon is not installed!')
             return {'CANCELLED'}
 
-        face = utilities.get_active_face()
-        if face and face.rig_logic_instance and face.head_mesh_object and face.head_rig_object:
-            instance = face.rig_logic_instance
+        head = utilities.get_active_head()
+        if head and head.rig_logic_instance and head.head_mesh_object and head.head_rig_object:
+            instance = head.rig_logic_instance
             dna_io_instance: DNAExporter = None # type: ignore
 
             # sync the spine bones with the body skeleton in the unreal blueprint
@@ -664,20 +671,20 @@ class SendToUnreal(bpy.types.Operator):
 
             # export a separate DNA file since we are only going to export bone 
             # transforms, mesh are sent across via FBX
-            dna_file = f'export/{face.rig_logic_instance.name}.dna'
-            if face.rig_logic_instance.output_method == 'calibrate':
+            dna_file = f'export/{head.rig_logic_instance.name}.dna'
+            if head.rig_logic_instance.output_method == 'calibrate':
                 dna_io_instance = DNACalibrator(
-                    instance=face.rig_logic_instance,
-                    linear_modifier=face.linear_modifier,
+                    instance=head.rig_logic_instance,
+                    linear_modifier=head.linear_modifier,
                     meshes=False,
                     vertex_colors=False,
                     bones=True,
                     file_name=dna_file
                 )              
-            elif face.rig_logic_instance.output_method == 'overwrite':
+            elif head.rig_logic_instance.output_method == 'overwrite':
                 dna_io_instance = DNAExporter(
-                    instance=face.rig_logic_instance,
-                    linear_modifier=face.linear_modifier,
+                    instance=head.rig_logic_instance,
+                    linear_modifier=head.linear_modifier,
                     meshes=False,
                     vertex_colors=False,
                     bones=True,
@@ -749,9 +756,9 @@ class ExportToDisk(bpy.types.Operator):
     bl_label = "Export to Disk"
 
     def execute(self, context):
-        face = utilities.get_active_face()
-        if face and face.rig_logic_instance:
-            instance = face.rig_logic_instance
+        head = utilities.get_active_head()
+        if head and head.rig_logic_instance:
+            instance = head.rig_logic_instance
             
             if not bpy.path.abspath(instance.output_folder_path) and not bpy.data.filepath:
                 self.report({'ERROR'}, 'File must be saved to use a relative path')
@@ -761,12 +768,12 @@ class ExportToDisk(bpy.types.Operator):
             if instance.output_method == 'calibrate':
                 dna_io_instance = DNACalibrator(
                     instance=instance,
-                    linear_modifier=face.linear_modifier
+                    linear_modifier=head.linear_modifier
                 )              
             elif instance.output_method == 'overwrite':
                 dna_io_instance = DNAExporter(
                     instance=instance,
-                    linear_modifier=face.linear_modifier
+                    linear_modifier=head.linear_modifier
                 )
 
             valid, title, message, fix = dna_io_instance.run()
@@ -813,9 +820,9 @@ class MirrorSelectedBones(bpy.types.Operator):
     bl_label = "Mirror Selected Bones"
 
     def execute(self, context):
-        face = utilities.get_active_face()
-        if face:
-            success, message = face.mirror_selected_bones()        
+        head = utilities.get_active_head()
+        if head:
+            success, message = head.mirror_selected_bones()        
             if not success:
                 self.report({'ERROR'}, message)
                 return {'CANCELLED'}
@@ -824,37 +831,6 @@ class MirrorSelectedBones(bpy.types.Operator):
     @classmethod
     def poll(cls, context):
         return callbacks.poll_head_rig_bone_selection(cls, context)
-    
-class PushBonesForwardAlongNormals(bpy.types.Operator):
-    """Pushes the selected bone positions forward along the mesh normals"""
-    bl_idname = "meta_human_dna.push_bones_forward_along_normals"
-    bl_label = "Push Bones Forward Along Normals"
-
-    def execute(self, context):
-        face = utilities.get_active_face()
-        if face:
-            face.push_selected_bones_along_mesh_normals(direction='forward')
-        return {'FINISHED'}
-    
-    @classmethod
-    def poll(cls, context):
-        return callbacks.poll_head_rig_bone_selection(cls, context)
-
-class PushBonesBackwardAlongNormals(bpy.types.Operator):
-    """Pushes the selected bone positions backward along the mesh normals"""
-    bl_idname = "meta_human_dna.push_bones_backward_along_normals"
-    bl_label = "Push Bones Backward Along Normals"
-
-    def execute(self, context):
-        face = utilities.get_active_face()
-        if face:
-            face.push_selected_bones_along_mesh_normals(direction='backward')
-        return {'FINISHED'}
-    
-    @classmethod
-    def poll(cls, context):
-        return callbacks.poll_head_rig_bone_selection(cls, context)
-
 
 class ShrinkWrapVertexGroup(bpy.types.Operator):
     """Shrink wraps the active vertex group on the head mesh using the shrink wrap modifier"""
@@ -862,9 +838,9 @@ class ShrinkWrapVertexGroup(bpy.types.Operator):
     bl_label = "Shrink Wrap Active Group"
 
     def execute(self, context):
-        face = utilities.get_active_face()
-        if face:
-            face.shrink_wrap_vertex_group()
+        head = utilities.get_active_head()
+        if head:
+            head.shrink_wrap_vertex_group()
         return {'FINISHED'}
     
 class AutoFitSelectedBones(bpy.types.Operator):
@@ -873,8 +849,8 @@ class AutoFitSelectedBones(bpy.types.Operator):
     bl_label = "Auto Fit Selected Bones"
 
     def execute(self, context):
-        face = utilities.get_active_face()
-        if face and face.head_mesh_object and face.head_rig_object:
+        head = utilities.get_active_head()
+        if head and head.head_mesh_object and head.head_rig_object:
             if bpy.context.mode != 'POSE': # type: ignore
                 self.report({'ERROR'}, 'You must be in pose mode')
                 return {'CANCELLED'}
@@ -884,14 +860,14 @@ class AutoFitSelectedBones(bpy.types.Operator):
                 return {'CANCELLED'}
             
             for pose_bone in bpy.context.selected_pose_bones: # type: ignore
-                if pose_bone.id_data != face.head_rig_object:
-                    self.report({'ERROR'}, f'The selected bone "{pose_bone.id_data.name}:{pose_bone.name}" is not associated with the rig logic instance "{face.rig_logic_instance.name}"')
+                if pose_bone.id_data != head.head_rig_object:
+                    self.report({'ERROR'}, f'The selected bone "{pose_bone.id_data.name}:{pose_bone.name}" is not associated with the rig logic instance "{head.rig_logic_instance.name}"')
                     return {'CANCELLED'}
             
             utilities.auto_fit_bones(
-                mesh_object=face.head_mesh_object, 
-                armature_object=face.head_rig_object,
-                dna_reader=face.dna_reader,
+                mesh_object=head.head_mesh_object, 
+                armature_object=head.head_rig_object,
+                dna_reader=head.dna_reader,
                 only_selected=True
             )
 
@@ -907,17 +883,17 @@ class RevertBoneTransformsToDna(bpy.types.Operator):
     bl_label = "Revert Bone Transforms to DNA"
 
     def execute(self, context):
-        face = utilities.get_active_face()
-        if face:
+        head = utilities.get_active_head()
+        if head:
             if bpy.context.mode != 'POSE': # type: ignore
                 self.report({'ERROR'}, 'Must be in pose mode')
                 return {'CANCELLED'}
             
-            if not face.rig_logic_instance.head_rig:
-                self.report({'ERROR'}, f'"{face.rig_logic_instance.name}" does not have a head rig assigned')
+            if not head.rig_logic_instance.head_rig:
+                self.report({'ERROR'}, f'"{head.rig_logic_instance.name}" does not have a head rig assigned')
                 return {'CANCELLED'}
 
-            face.revert_bone_transforms_to_dna()
+            head.revert_bone_transforms_to_dna()
         return {'FINISHED'}
     
     @classmethod
@@ -997,6 +973,9 @@ class MetaHumanDnaReportError(ShapeKeyOperatorBase):
 
     def invoke(self, context, event):
         wm = context.window_manager # type: ignore
+        if not wm:
+            return
+
         fix = wm.meta_human_dna.errors.get(self.title, {}).get('fix', None) # type: ignore
         return wm.invoke_props_dialog(
             self,
@@ -1006,6 +985,9 @@ class MetaHumanDnaReportError(ShapeKeyOperatorBase):
         )
 
     def draw(self, context):
+        if not self.layout:
+            return
+        
         for line in self.title.split('\n'):
             row = self.layout.row()
             row.scale_y = 1.5
@@ -1030,6 +1012,9 @@ class MetricsCollectionConsent(bpy.types.Operator):
 
     def invoke(self, context, event):
         wm = context.window_manager # type: ignore
+        if not wm:
+            return
+        
         preferences = context.preferences.addons[ToolInfo.NAME].preferences # type: ignore
         current_timestamp = datetime.now().timestamp()
 
@@ -1055,9 +1040,11 @@ class MetricsCollectionConsent(bpy.types.Operator):
         preferences.next_metrics_consent_timestamp = (datetime.now() + timedelta(days=30)).timestamp() # type: ignore
         preferences.metrics_collection = False # type: ignore
         bpy.ops.meta_human_dna.force_evaluate() # type: ignore
-        return {'CANCELLED'}
 
     def draw(self, context):
+        if not self.layout:
+            return
+        
         row = self.layout.row()
         row.label(text="We collect anonymous metrics and bug reports to help improve the MetaHuman DNA addon.")
         row = self.layout.row()
@@ -1129,9 +1116,9 @@ class ReImportThisShapeKey(ShapeKeyOperatorBase):
     shape_key_name: bpy.props.StringProperty(name="Shape Key Name") # type: ignore
 
     def execute(self, context):
-        face = utilities.get_active_face()
-        if face and face.rig_logic_instance:
-            instance = face.rig_logic_instance
+        head = utilities.get_active_head()
+        if head and head.rig_logic_instance:
+            instance = head.rig_logic_instance
             result = self.validate(context, instance)
             if not result:
                 return {'CANCELLED'}
@@ -1156,7 +1143,7 @@ class ReImportThisShapeKey(ShapeKeyOperatorBase):
                 name=short_name,
                 prefix=f'{mesh_dna_name}__',
                 is_neutral=instance.generate_neutral_shapes,
-                linear_modifier=face.linear_modifier,
+                linear_modifier=head.linear_modifier,
                 delta_threshold=0.0001
             )
             mesh_object.show_only_shape_key = False
@@ -1328,9 +1315,11 @@ class DuplicateRigLogicInstance(bpy.types.Operator):
         return callbacks.get_active_rig_logic() is not None
     
     def draw(self, context):
-        layout = self.layout
-        layout.prop(self, 'new_name')
-        layout.prop(self, 'new_folder')
+        if not self.layout:
+            return
+
+        self.layout.prop(self, 'new_name')
+        self.layout.prop(self, 'new_folder')
 
 
 class AddRigLogicTextureNode(bpy.types.Operator):
